requirements.txt
beautifulsoup4>=4.12.0
httpx>=0.24.0
tenacity>=8.2.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
openai>=1.3.0
supabase>=1.2.0
langchain>=0.1.0
langchain_text_splitters>=0.0.1
astra-db-python>=0.4.0
sentence-transformers>=2.2.2 